+++ dev_tools/test.ps1
@@ -7,11 +7,6 @@
 #
 # SPDX-License-Identifier: EPL-2.0
 ##########################
-
-# Launch the triplestore
-# Then install, lanuch and test the adapter
-
-
 param (
     [switch]$t,
     [switch]$l,
@@ -30,121 +25,50 @@ $USAGE="   Usage: $PSCommandPath [-t|-h|-l]
 "
 
 # duration for polling (via curl)
-$SLEEP=1
+$SLEEP=3
 
 $USRPATH=$(pwd)         # get the call directory
 $ROOTDIR=$PSScriptRoot  # get the script directory
 cd $ROOTDIR             # move to the script directory
 cd ..
 $ADAPTER_ROOT_DIR=$(pwd) # get the adapter root directory
-cd $ROOTDIR             # move back to the script directory
+cd $USRPATH              # move back to the user
 
 $PIDS_TO_KILL=@()
 
-function print_help() {
-    echo "$HELP"
-    echo "$USAGE"
-    exit 0
-}
-
-# $1 name of the invalid arg
-function invalid_arg() {
-    param (
-        $1
-    )
-    echo "`n   Invalid argument: ${1} `n"
-    echo "$USAGE"
-    exit 1
-}
-
-# polls an address using curl until the request returns True (i.e. the address responds)
-# $1 ... URL for curl to poll
-function curl_poll
-{
-    param (
-        $1
-    )
-    $curl_ret=42
-    while ( $curl_ret -ne $true )
-    {
-        Start-Sleep -Seconds $SLEEP
-        $curl_ret = $false
-        Write-Host -NoNewline  "."
-        try{
-            Invoke-WebRequest -Uri $1 -ErrorAction Ignore 2> $null > $null
-        } catch {
-        }
-        $curl_ret=$?
-    }
-    echo ""
-}
-
-
-# Check that all required configuration files exist.
-# Outputs an error message and exits the script if 
-# a conf file is missing.
-function checkConfFiles()
-{
-    if (!(Test-Path "$ADAPTER_ROOT_DIR/analysis/conf/VeriFitAnalysis.properties" -PathType Leaf)) {
-        echo 'ERROR: Configuration file "'$ADAPTER_ROOT_DIR'"/analysis/conf/VeriFitAnalysis.properties" not found.'
-        exit 1
-    }
-    if (!(Test-Path "$ADAPTER_ROOT_DIR/compilation/conf/VeriFitCompilation.properties" -PathType Leaf)) {
-        echo 'ERROR: Configuration file "'$ADAPTER_ROOT_DIR'"/compilation/conf/VeriFitCompilation.properties" not found.'
-        exit 1
-    }
-    if (!(Test-Path "$ADAPTER_ROOT_DIR/sparql_triplestore/start.ini" -PathType Leaf)) {
-        echo 'ERROR: Configuration file "'$ADAPTER_ROOT_DIR'"/sparql_triplestore/start.ini" not found.'
-        exit 1
-    }
-}
-
-# Kills a process and its children recursively
-# $1 ... pid of the process
-function KillWithChildren {
-    Param(
-        $1
-    )
-    Get-WmiObject win32_process | where {$_.ParentProcessId -eq $1} | ForEach { KillWithChildren $_.ProcessId }
-    Stop-Process $1 2> $null
-}
-
-# kills all children of this process using the $PIDS_TO_KILL variable
-function killChildren {
-    echo "`nShutting down..."
-    foreach ($i in $PIDS_TO_KILL) {
-        KillWithChildren $i
-    }
-    echo "All done."
-}
+# source shared utils
+. $ADAPTER_ROOT_DIR/dev_tools/shared.ps1
 
 #
 # main
 #
 if ($args.length -ne 0) {
-    invalid_arg $args[0]
+    invalid_arg $args[0] $USAGE
 }
 if ($h) {
-    print_help
+    print_help $HELP $USAGE
 }
 
-# make sure configuration files exist
-checkConfFiles
+
+echo "`nNOTE:`n  MAKE SURE THAT THE ADAPTER IS RUNNING ON THE DEFAULT PORTS!`n  THE TESTSUITE CURRENTLY HAS PORTS HARDCODED INSIDE IT`n" # TODO
+
+echo "`nNOTE2:`n  If this script gets killed prematurely, then multiple minimized`n  powershell windows stay open and need to be closed by hand`n" # TODO
 
 # lookup analysis adapter config
-$analysis_host=$(cat $ADAPTER_ROOT_DIR/analysis/conf/VeriFitAnalysis.properties | Select-String -Pattern "^ *adapter_host=") -replace "^ *adapter_host=", "" -replace "/$", ""
-$analysis_port=$(cat $ADAPTER_ROOT_DIR/analysis/conf/VeriFitAnalysis.properties | Select-String -Pattern "^ *adapter_port=") -replace "^ *adapter_port=", ""
-$analysis_url="${analysis_host}:${analysis_port}/analysis/"
+$analysis_url = lookupAnalysisURL "$ADAPTER_ROOT_DIR"
 
 
 if ( ! $l ) {
     echo "Booting up the Universal Analysis Adapter"
     $process = Start-Process -WindowStyle Minimized powershell.exe "(Get-Host).ui.RawUI.WindowTitle='Universal Analysis Adapter'; $ADAPTER_ROOT_DIR/run_all.ps1" -passthru
     $PIDS_TO_KILL = $PIDS_TO_KILL + $process.id
-    curl_poll $analysis_url       # poll the analysis adapter because that one starts last in the run script
+    $ret = curl_poll $analysis_url $SLEEP  # poll the analysis adapter because that one starts last in the run script
     echo "Adapter up and running`n"
 } else {
-    echo "Skipping Adapter boot. Adapter expected to be running already." 
+    echo "Skipping Adapter boot. Adapter expected to be running already."
+
+    # make sure configuration files exist
+    checkConfFiles "$ADAPTER_ROOT_DIR"
 }
 
 
@@ -182,9 +106,7 @@ if ($t) {
 }
 
 # shutdown the adapter
-killChildren
-
-cd $USRPATH
+killAllWithChildren $PIDS_TO_KILL
 
 # return non-zero if there were failed tests
 if ( $compilationRes -ne $true -Or $analysisRes -ne $true -Or $analysisToolsRes -ne $true) {