+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ### v2.1.0
-- TODO
+- added a ./conf directory and a ./conf_example directory
+  - the ./conf directory is where users place all their custom configuration files
+  - both the build script and the run script will then load the configuration from
+    the ./conf directory and copy it into appropriate places
+  - see the ./conf_example to learn what configuration files to place in the ./conf dir
+- reworked scripts to used shared resources
+- moved the jetty distribution into the ./lib directory and then use jetty base in ./sparql_triplestore
+- changed built in tool loading to be loaded the same way as custom tools
+- changed powershell exec in the analysis adapter to print exceptions properly
+- ... 
 
 ### v2.0.0
 - added a new common input parameter "outputFilter" to the analysis adapter